Fixed Login-Page Frontend issues
Fixed the HackMerced logo being too big at lower resolutions, added a paragraph to login-resetpassword page, fixed the auth login buttons on the login page, and increased the width of the submit buttons on login page and login-resetpassword page. We still need to link the 'Create an Account' button to the sign-up page, and link the three auth icons to the correct links as well as work on the backend for them.